chore(bot): better synchronization of bot save interactions #22
