+++ bot/middlewares/index.js
@@ -1,4 +1,3 @@
 export { default as handleMessage } from './handle-message'
-export { default as saveReceivedInteraction } from './save-received-interaction'
 export { default as sendMassMessage } from './send-mass-message'
 export { default as verifyValidationToken } from './verify-validation-token'
